requirements.txt
Click==7.0
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
gunicorn==20.0.4
imageio==2.6.1
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.2.1
pkg-resources==0.0.0
pyparsing==2.4.5
pytesseract==0.3.0
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.3.3
six==1.13.0
Werkzeug==0.16.0
